Removing all components from Main and split them into Layout, NavBar and Footer components
