+++ src/components/Layout/styled/index.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components'
+
+export const Container = styled.div`
+  flex-grow: 1;
+  margin: auto;
+  width: 80%;
+`
+
+export const Wrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+`
