--- src/components/Main/Menu/index.js
@@ -1,1 +1,0 @@
-export { default as Menu } from './Menu'
