--- src/components/Main/index.js
@@ -1,2 +1,0 @@
-export { default as Main } from './Main'
-export * from './styled'
