--- src/components/Main/styled/Content.js
@@ -1,9 +1,0 @@
-import styled from 'styled-components'
-
-const Content = styled.div`
-  flex-grow: 1;
-  margin: auto;
-  width: 80%;
-`
-
-export default Content
