--- src/components/Main/styled/Li.js
@@ -1,8 +1,0 @@
-import styled from 'styled-components'
-
-const Li = styled.li`
-  padding: 0 .5rem;
-  width: auto;
-`
-
-export default Li
