--- src/components/Main/styled/NavItems.js
@@ -1,12 +1,0 @@
-import styled from 'styled-components'
-
-const NavItems = styled.ul`
-  align-items: center;
-  display: flex;
-  flex-flow: row wrap;
-  justify-content: space-between;
-  margin: auto;
-  padding: 0;
-`
-
-export default NavItems
