--- src/components/Main/styled/Page.js
@@ -1,9 +1,0 @@
-import styled from 'styled-components'
-
-const Page = styled.div`
-  display: flex;
-  flex-direction: column;
-  min-height: 100vh;
-`
-
-export default Page
