--- src/components/Main/styled/index.js
@@ -1,9 +1,0 @@
-export * from './Brand'
-export { default as Content } from './Content'
-export { default as Footer } from './Footer'
-export { default as Li } from './Li'
-export { default as LinkItem } from './LinkItem'
-export { default as Nav } from './Nav'
-export { default as NavItems } from './NavItems'
-export { default as NavContainer } from './NavContainer'
-export { default as Page } from './Page'
